Remove converter and unnecessary table constraint conversion
