+++ floor_generator/lib/processor/query_analyzer/type_checker.dart
@@ -8,8 +8,6 @@ import 'package:source_gen/source_gen.dart';
 import 'package:sqlparser/sqlparser.dart';
 import 'package:floor_generator/misc/type_utils.dart';
 
-import 'converter.dart';
-
 //TODO TypeConverters!!
 
 void assertMatchingTypes(List<Field> fields,
